adds edit button, some styling

+++ src/App.js
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
+import {
+  FaCircle, FaEdit, FaTrashAlt,
+} from 'react-icons/fa';
 import config from './config';
-import CreateButton from './components/CreateButton';
-import Title from './components/Title';
 import ColorPalette from './components/ColorPalette';
 
 const APP_DATA_KEY = 'appData';
@@ -22,6 +23,7 @@ const getRandomColor = () => ({
     start: config.lightness.min,
     range: 100,
   },
+  editing: false,
 });
 
 function getInitialColors() {
@@ -49,10 +51,37 @@ function App() {
   }
 
   function handleDelete(idx) {
-    const newColors = [...colors.slice(0, idx), ...colors.slice(idx + 1)];
-    localStorage.setItem(APP_DATA_KEY, JSON.stringify(newColors));
-    setColors(newColors);
+    const deleteItem = window.confirm(`Are you sure you want to delete ${colors[idx].name}?`);
+    if (deleteItem) {
+      const newColors = [...colors.slice(0, idx), ...colors.slice(idx + 1)];
+      localStorage.setItem(APP_DATA_KEY, JSON.stringify(newColors));
+      setColors(newColors);
+    }
   }
+
+  const styles = {
+    listDisc: `
+    w-1/12
+    text-2xl
+    my-auto
+    mr-2
+    `,
+
+    listItem: `
+    flex
+    `,
+
+    listItemText: `
+    w-11/12
+    `,
+
+    inlineIcon: `
+    inline
+    text-xl
+    text-gray-700
+    mb-1
+    `,
+  };
 
   return (
     <AppContext.Provider
@@ -63,10 +92,64 @@ function App() {
         handleDelete,
       }}
     >
-      <div className="max-w-3xl mx-auto relative font-mono">
-        <Title />
+      <div className="max-w-3xl px-2 md:px-4 mx-auto relative">
+        <h1 className="text-5xl font-medium leading-none md:text-6xl md:mt-16 block">Color Generator</h1>
+        <ul className="flex flex-col space-y-6 mt-12 text-sm md:text-base text-gray-800">
+          <li className={styles.listItem}>
+            <FaCircle className={`${styles.listDisc} text-green-500`} />
+            <div className={styles.listItemText}>
+              Click the
+              {' '}
+              <FaEdit
+                title="Pencil icon"
+                className={styles.inlineIcon}
+              />
+              {' '}
+              to edit and the
+              {' '}
+              <FaTrashAlt
+                title="Trashcan icon"
+                className={styles.inlineIcon}
+              />
+              {' '}
+              to delete a color.
+
+            </div>
+          </li>
+          <li className={styles.listItem}>
+            <FaCircle className={`${styles.listDisc} text-blue-500`} />
+            <div className={styles.listItemText}>
+              Click a color shade to copy its HSL value.
+
+            </div>
+          </li>
+          <li className={styles.listItem}>
+            <FaCircle className={`${styles.listDisc} text-red-500`} />
+            <div className={styles.listItemText}>
+              Your work is automatically saved to your browser&apos;s local storage.
+            </div>
+          </li>
+        </ul>
         <ColorPalette />
-        <CreateButton />
+        {!colors.length && (
+        <h2 className="text-4xl text-gray-800">
+          <span
+            role="img"
+            aria-label="pointing down"
+            className="mr-4 text-5xl"
+          >
+            👇
+          </span>
+          Click to get started!
+        </h2>
+        )}
+        <button
+          type="button"
+          onClick={handleCreate}
+          className="bg-green-500 rounded text-xl mt-12 mb-12 hover:bg-green-600 text-green-100 px-5 py-2 focus:outline-none"
+        >
+          Add Color
+        </button>
       </div>
     </AppContext.Provider>
   );
